deprecate hardware info interval parameter
The hardware info message will only be sent after onboarding and after
reboots. The hardware health message will continue to be sent
periodically as before. Therefore there is no need for a separate
interval parameter for hardware info.
