+++ pkg/pillar/cmd/zedagent/hardwareinfo.go
@@ -13,7 +13,6 @@ import (
 	"github.com/lf-edge/eve-api/go/info"
 	"github.com/lf-edge/eve/pkg/pillar/agentlog"
 	"github.com/lf-edge/eve/pkg/pillar/hardware"
-	"github.com/lf-edge/eve/pkg/pillar/types"
 	"google.golang.org/protobuf/proto"
 	"google.golang.org/protobuf/types/known/timestamppb"
 )
@@ -21,13 +20,10 @@ import (
 func hardwareInfoTask(ctxPtr *zedagentContext, triggerHwInfo <-chan destinationBitset) {
 	wdName := agentName + "hwinfo"
 
-	ticker := time.NewTicker(10 * time.Second)
-
 	stillRunning := time.NewTicker(30 * time.Second)
 	ctxPtr.ps.StillRunning(wdName, warningTime, errorTime)
 	ctxPtr.ps.RegisterFileWatchdog(wdName)
 
-	ts := time.Now()
 	for {
 		select {
 		case dest := <-triggerHwInfo:
@@ -39,20 +35,10 @@ func hardwareInfoTask(ctxPtr *zedagentContext, triggerHwInfo <-chan destinationB
 			log.Function("HardwareInfoTask done with message")
 			ctxPtr.ps.CheckMaxTimeTopic(wdName, "PublishHardwareInfo", start,
 				warningTime, errorTime)
-		case <-ticker.C: // even if no trigger comes, publish hardware info periodically
-			if time.Now().Sub(ts) > getHardwareInfoInterval(ctxPtr) {
-				ts = time.Now()
-				triggerPublishHwInfo(ctxPtr)
-			}
 		case <-stillRunning.C:
 		}
 		ctxPtr.ps.StillRunning(wdName, warningTime, errorTime)
 	}
-}
-
-func getHardwareInfoInterval(ctx *zedagentContext) time.Duration {
-	interval := ctx.globalConfig.GlobalValueInt(types.HardwareInfoInterval)
-	return time.Duration(interval) * time.Second
 }
 
 func triggerPublishHwInfoToDest(ctxPtr *zedagentContext, dest destinationBitset) {
@@ -141,26 +127,6 @@ func PublishHardwareInfoToZedCloud(ctx *zedagentContext, dest destinationBitset)
 
 	queueInfoToDest(ctx, dest, hwInfoKey, buf, bailOnHTTPErr, false, false,
 		info.ZInfoTypes_ZiHardware)
-}
-
-func getSmartAttr(diskData []*types.DAttrTable) []*info.SmartAttr {
-	attrResults := []*info.SmartAttr{} // Store pointers instead of structs
-
-	for _, attr := range diskData {
-		attrResult := &info.SmartAttr{ // Allocate on heap
-			Id:            uint32(attr.ID),
-			AttributeName: attr.AttributeName,
-			RawValue:      uint64(attr.RawValue),
-			Thresh:        uint64(attr.Threshold),
-			Worst:         uint64(attr.Worst),
-			Value:         uint64(attr.Value),
-			Type:          attr.Type,
-		}
-
-		attrResults = append(attrResults, attrResult) // Append pointer
-	}
-
-	return attrResults
 }
 
 func getKernelVersion() string {
